+++ frontend/Sentinel/package.json
@@ -11,6 +11,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@hookform/resolvers": "^5.1.1",
     "@react-native-async-storage/async-storage": "2.1.2",
     "@react-native-community/netinfo": "^11.4.1",
     "@react-navigation/bottom-tabs": "^7.3.14",
@@ -43,6 +44,7 @@
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
     "react": "19.0.0",
+    "react-hook-form": "^7.58.1",
     "react-native": "0.79.3",
     "react-native-dotenv": "^3.4.11",
     "react-native-gesture-handler": "~2.24.0",
@@ -56,7 +58,8 @@
     "redux-persist": "^6.0.0",
     "root": "tanstack/react-query",
     "victory-native": "^36.9.2",
-    "yup": "^1.6.1"
+    "yup": "^1.6.1",
+    "zod": "^3.25.67"
   },
   "devDependencies": {
     "@babel/core": "^7.27.4",
